obchod s npc , uprava dialogu , boje , liecitelský npc ,

+++ BakalarkaTahoveRPG/AkciaLiecenie.h
@@ -5,13 +5,13 @@
 class AkciaLiecenie:public Akcia
 {
 public:
-	AkciaLiecenie(std::string meno, std::string obrazok, int casCastenia, int cooldown, int trvanie, std::string popis, int mana, AkciaTyp typ, int paZakladnyHeal);
+	AkciaLiecenie(std::string meno, std::string obrazok, int casCastenia, int cooldown, int trvanie, std::string popis, int mana, AkciaTyp typ, double paZakladnyHeal);
 	~AkciaLiecenie();
 	std::string vykonajSa(Statistika* statHrac, Statistika* statNepriatel, sf::Time aktCas);
 	int minLiecenie(Statistika* stat);
 	int maxLiecenie(Statistika* stat);
 private:
-	int zakladnyHeal;
+	double zakladnyHeal;
 };
 
 #endif
